+++ app/src/main/java/com/example/week5/UIcomponents/BasketItem.kt
@@ -1,0 +1,62 @@
+package com.example.week5.UIcomponents
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import com.example.week5.Data.Product
+
+/**
+ * Карточка продукта для отображения в [BasketScreen]
+ *
+ * @param product конкретный продукт для отображения
+ * @param onDelete лябда-функция для удаления текущего продукта из [BasketScreen]
+ */
+@Composable
+fun BasketItem(
+    product: Product,
+    onDelete: () -> Unit
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(16.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        AsyncImage(
+            model = product.photoUri,
+            contentDescription = product.title,
+            modifier = Modifier.size(64.dp),
+            contentScale = ContentScale.Crop
+        )
+        Column(
+            modifier = Modifier
+                .weight(1f)
+                .padding(horizontal = 16.dp)
+        ) {
+            Text(product.title)
+            Text(
+                text = product.content.take(30) + if (product.content.length > 30) "..." else "",
+                style = MaterialTheme.typography.bodySmall
+            )
+        }
+        IconButton(onClick = onDelete) {
+            Icon(Icons.Filled.Delete, contentDescription = "УдалитьКнопка")
+        }
+    }
+    HorizontalDivider()
+}
